requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
streamlit-autorefresh>=1.0.1
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
